Final moves
Moved all the packages in the producer folder into internal and changed all imports. All tests seem to be passing
