+++ cmd/producer/main/main.go
@@ -20,7 +20,7 @@ import (
 	"github.com/SIGBlockchain/project_aurum/internal/blockchain"
 	"github.com/SIGBlockchain/project_aurum/internal/constants"
 	"github.com/SIGBlockchain/project_aurum/internal/genesis"
-	"github.com/SIGBlockchain/project_aurum/internal/producer/src/producer"
+	"github.com/SIGBlockchain/project_aurum/internal/producer"
 
 	"github.com/pborman/getopt"
 )
