+++ internal/block/block.go
@@ -10,8 +10,8 @@ import (
 	"reflect"
 	"time"
 
-	"github.com/SIGBlockchain/project_aurum/internal/producer/src/contracts"
-	"github.com/SIGBlockchain/project_aurum/internal/producer/src/hashing"
+	"github.com/SIGBlockchain/project_aurum/internal/contracts"
+	"github.com/SIGBlockchain/project_aurum/internal/hashing"
 )
 
 type BlockHeader struct {
