+++ internal/blockchain/blockchain_test.go
@@ -18,14 +18,14 @@ import (
 	_ "github.com/mattn/go-sqlite3"
 
 	"github.com/SIGBlockchain/project_aurum/internal/accountstable"
+	block "github.com/SIGBlockchain/project_aurum/internal/block"
 	"github.com/SIGBlockchain/project_aurum/internal/constants"
+	"github.com/SIGBlockchain/project_aurum/internal/contracts"
 	"github.com/SIGBlockchain/project_aurum/internal/genesis"
-	block "github.com/SIGBlockchain/project_aurum/internal/producer/src/block"
-	"github.com/SIGBlockchain/project_aurum/internal/producer/src/contracts"
-	"github.com/SIGBlockchain/project_aurum/internal/producer/src/hashing"
-	"github.com/SIGBlockchain/project_aurum/internal/producer/src/validation"
+	"github.com/SIGBlockchain/project_aurum/internal/hashing"
 	"github.com/SIGBlockchain/project_aurum/internal/publickey"
 	"github.com/SIGBlockchain/project_aurum/internal/sqlstatements"
+	"github.com/SIGBlockchain/project_aurum/internal/validation"
 )
 
 func setUp(filename string, database string) {
