+++ internal/client/src/client/client_test.go
@@ -10,22 +10,20 @@ import (
 	"encoding/pem"
 	"fmt"
 	"io/ioutil"
-	"log"
 	"math/rand"
 	"net"
 	"os"
 	"reflect"
 	"testing"
-	"time"
 
 	"github.com/SIGBlockchain/project_aurum/internal/constants"
 
 	"github.com/SIGBlockchain/project_aurum/internal/accountstable"
-	"github.com/SIGBlockchain/project_aurum/internal/producer/src/hashing"
+	"github.com/SIGBlockchain/project_aurum/internal/hashing"
 	"github.com/SIGBlockchain/project_aurum/internal/privatekey"
 	"github.com/SIGBlockchain/project_aurum/internal/publickey"
 
-	producer "github.com/SIGBlockchain/project_aurum/internal/producer/src/producer"
+	producer "github.com/SIGBlockchain/project_aurum/internal/producer"
 )
 
 // Test will fail in airplane mode, or just remove wireless connection.
@@ -50,36 +48,6 @@ func TestGetUserInput(t *testing.T) {
 	if user_input != "TEST" {
 		t.Errorf("User Input Check Failed.")
 	}
-}
-
-// Test send to producer with small max length message for one send
-func TestSendToProducer(t *testing.T) {
-	sz := 1024
-	testbuf := make([]byte, sz)
-	for i, _ := range testbuf {
-		testbuf[i] = 1
-	}
-	addr := "localhost:8080"
-	ln, err := net.Listen("tcp", addr)
-	var buffer bytes.Buffer
-	bp := producer.BlockProducer{
-		Server:        ln,
-		NewConnection: make(chan net.Conn, 128),
-		Logger:        log.New(&buffer, "LOG:", log.Ldate),
-	}
-	go bp.AcceptConnections()
-	time.Sleep(1)
-	if err != nil {
-		t.Errorf("Failed to set up listener")
-	}
-	n, err := SendToProducer(testbuf, addr)
-	if err != nil {
-		t.Errorf("Failed to send to producer")
-	}
-	if n != sz {
-		t.Errorf("Did not write all bytes to connection")
-	}
-	ln.Close()
 }
 
 func TestSetupWallet(t *testing.T) {
