+++ internal/contracts/contracts.go
@@ -10,7 +10,7 @@ import (
 	"fmt"
 	"reflect"
 
-	"github.com/SIGBlockchain/project_aurum/internal/producer/src/hashing"
+	"github.com/SIGBlockchain/project_aurum/internal/hashing"
 	"github.com/SIGBlockchain/project_aurum/internal/publickey"
 )
