+++ internal/contracts/contracts_test.go
@@ -12,7 +12,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/SIGBlockchain/project_aurum/internal/producer/src/hashing"
+	"github.com/SIGBlockchain/project_aurum/internal/hashing"
 	"github.com/SIGBlockchain/project_aurum/internal/publickey"
 )
