+++ internal/hashing/hashing.go
