+++ internal/requests/requests.go
@@ -7,8 +7,8 @@ import (
 	"errors"
 	"net/http"
 
+	"github.com/SIGBlockchain/project_aurum/internal/contracts"
 	"github.com/SIGBlockchain/project_aurum/internal/endpoints"
-	"github.com/SIGBlockchain/project_aurum/internal/producer/src/contracts"
 	"github.com/SIGBlockchain/project_aurum/internal/publickey"
 )
